Add files via upload
Big Update, had to change from a WindowManager popup to a panel, as the eyedropper could not be used from a popup window. Makes sense, also added asked for features like Overriding Selection to ensure a cut. All the options from bisect were taken over and the options to cut off(destuctive) mode added. The UI also reflects the active object and does not show itself if the selected object is no Mesh, like the camera or lights.